pyproject.toml
[tool.poetry]
name = "squirro_code"
version = "0.1.0"
description = ""
authors = ["AsgrimS <jan.semik@gmail.com>"]
readme = "README.md"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
motor = "^3.0.0"
pydantic = {extras = ["dotenv"], version = "<1.10.0"}
spacy = "^3.4.1"
fastapi-cache2 = "^0.1.9"

[tool.poetry.dependencies.en_core_web_sm]
url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0.tar.gz"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
flake8-isort = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
